+++ CHANGELOG.md
@@ -2,5 +2,12 @@
 mdr1986x-small-lib changelog
 
 ## [Unreleased]
+### Added
+- Added HSE and clock source type parameters in SYS_ClkInit for flexible clock
+setup. Changed version of system.c to 0.4.
 
-## [0.1]
+## [0.1] - 2019-06-25
+### Added
+- Initial lib files
+- This CHANGELOG file
+